+++ web/src/app/admin/connector/[ccPairId]/stage-metrics/interfaces.ts
@@ -1,0 +1,1 @@
+export type SortMode = "pipeline" | "time-taken";
